test: split parameter service handler tests
